% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyDistanceRule.R
\name{applyDistanceRule}
\alias{applyDistanceRule}
\title{A function for finding the closest vector to a centroid}
\usage{
applyDistanceRule(v, m)
}
\arguments{
\item{v}{a rule-based vector for comparison.}

\item{m}{a matrix-like object of centroid vectors}
}
\description{
Compares a matrix of of centroid vectors to a rule based vector
and returns the label (i.e. index) of the closest class
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
