% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearExperiment.R
\name{clearExperiment}
\alias{clearExperiment}
\title{A convenience function for running hypotheses}
\usage{
clearExperiment(hypothesis)
}
\arguments{
\item{hypothesis}{specifies the research question or hypothesis}
}
\description{
A convenience function for running hypotheses
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{loading}
\keyword{package}
