% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusionMatrix.R
\name{confusionMatrix}
\alias{confusionMatrix}
\title{A confusion matrix object}
\usage{
confusionMatrix(predictions, observed)
}
\arguments{
\item{predictions}{a vector of predicted states (e.g. 0 or 1)}

\item{observed}{a vector of observed or actual states}
}
\description{
This function defines a class of confusion matrix. It takes a prediction
vector and an observed vector as inputs, and returns a list like object
containing the confusion matrix and the relevant indices of accuracy already
pre-computed.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
