% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crmetric.R
\name{crmetric}
\alias{crmetric}
\title{Implements a crmetric class}
\usage{
crmetric(metric, careless, crModel)
}
\arguments{
\item{metric}{the metric in question}

\item{careless}{a binary vector indicating whether each respondent is careless}

\item{crModel}{a factor containing the CR model each careless respondent was drawn from}
}
\description{
This class returns a crmetric object
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
