% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crModel.R
\name{crModel}
\alias{crModel}
\title{A careless responding model object}
\usage{
crModel(label, mu, sigma)
}
\arguments{
\item{label}{the name applied to this type of careless responding}

\item{mu}{the hypothetical mean occurance in a sample}

\item{sigma}{the hypothetical standard deviation of occurance in a sample}
}
\description{
This function defines a model of careless responder. Objects using this class
are initialized with the following parameters
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{data}
\keyword{model}
