% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataModel.R
\name{dataModel}
\alias{dataModel}
\title{A data model object}
\usage{
dataModel(factorCorrelations, itemParameters, nchar)
}
\arguments{
\item{itemParameters}{a matrix of IRT item parameters}

\item{nchar}{the number of characters to retain from the factor labels}

\item{factorCorllations}{a matrix of factor correlations}
}
\description{
This function defines a class of data model. It takes a matrix of factor
correlations and a matrix of item parameters as inputs, then computes the
model slopes then returns a dataModel object which can be used as an input
for data simulation.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{data}
\keyword{model}
