% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchClassifierEval.R
\name{dispatchClassifierEval}
\alias{dispatchClassifierEval}
\title{An interface for testing classification methods}
\usage{
dispatchClassifierEval(x, method, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}

\item{method}{a string specifying the classification method to evaluate}
}
\description{
Takes a dataframe and method label as parameters, then evaluates the 
performance of the specified classification method, returning a
confusion_matrix object that contains the common classifier metrics.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
