% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchH1.R
\name{dispatchH1}
\alias{dispatchH1}
\title{Dispatcher for Hypothesis 1}
\usage{
dispatchH1(x, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}

\item{...}{any additional arguments}
}
\description{
Takes a dataframe as an argument and returns the result of hypothesis 2
for that dataframe
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
