% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchHypothesisEval.R
\name{dispatchHypothesisEval}
\alias{dispatchHypothesisEval}
\title{A hypothesis dispatcher}
\usage{
dispatchHypothesisEval(x, hypothesis, ...)
}
\arguments{
\item{x}{a dataframe for analysis}

\item{...}{any additional parameters}
}
\description{
A hypothesis dispatcher
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
