% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchRQ1.R
\name{dispatchRQ1}
\alias{dispatchRQ1}
\title{An interface for optimizing classification methods}
\usage{
dispatchRQ1(x, what, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}

\item{what}{a string specifying the index to optimize}

\item{...}{any additional arguments}
}
\description{
Takes a dataframe and a label for what you're optimizing as parameters, 
then it optimizes the specified classifier.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
