% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchRQ2.R
\name{dispatchRQ2}
\alias{dispatchRQ2}
\title{A hypothesis dispatcher}
\usage{
dispatchRQ2(x, ...)
}
\arguments{
\item{x}{a dataframe for analysis}

\item{...}{any additional parameters}
}
\description{
A hypothesis dispatcher
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
