% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispatchRQ3.R
\name{dispatchRQ3}
\alias{dispatchRQ3}
\title{A dispatcher for Research Question 3}
\usage{
dispatchRQ3(x, ...)
}
\arguments{
\item{x}{a dataframe for analysis}

\item{...}{any additional parameters}
}
\description{
A dispatcher for Research Question 3
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
