% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateHypothesis1.R
\name{evaluateHypothesis1}
\alias{evaluateHypothesis1}
\title{Evaluates Hypothesis 1 on a single dataset}
\usage{
evaluateHypothesis1(x, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}

\item{...}{other parameters as necessary}
}
\description{
Takes a dataframe and dots as parameters
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
