% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluateHypothesis2.R
\name{evaluateHypothesis2}
\alias{evaluateHypothesis2}
\title{Evaluates Hypothesis 2 on a single dataset}
\usage{
evaluateHypothesis2(x, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}

\item{...}{other parameters as necessary}
}
\description{
Takes a dataframe and dots as parameters
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
