% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evenoddClassifier.R
\name{evenoddClassifier}
\alias{evenoddClassifier}
\title{Implements the evenodd classifier evaluation}
\usage{
evenoddClassifier(x, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe as an argument and evaluates the performance of evenodd
as a classifier for that dataset.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
