% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCentered.R
\name{genCentered}
\alias{genCentered}
\title{Simulates a centered careless respondent}
\usage{
genCentered(x, floor = 1, ceiling = 7)
}
\arguments{
\item{x}{the length of the span of careless responses}

\item{floor}{the lowest possible value for a response}

\item{ceilingthe}{highest possible value for a response}
}
\description{
Given a vector of survey responses, an insert, and a number of items,
this function simulates a centered careless respondent.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
