% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSkew.R
\name{genSkew}
\alias{genSkew}
\title{Simulates a skewed careless respondent}
\usage{
genSkew(x, mu = 5.75, sigma = 1.25, floor = 1, ceiling = 7)
}
\arguments{
\item{x}{the length of the span of careless responses}

\item{mu}{the mean of the distribution from from which to draw the skewed respondent}

\item{sigma}{the standard deviation of the distribution from which to draw the skewed respondent}

\item{floor}{the lowest possible value for a response}

\item{the}{highest possible value for a response}
}
\description{
Given a vector of survey responses, an insert, and a number of items,
this function simulates a skewed careless respondent.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
