% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getProportionFromModel.R
\name{getProportionFromModel}
\alias{getProportionFromModel}
\title{A careless responding model object}
\usage{
getProportionFromModel(m, n)
}
\arguments{
\item{m}{a crModel object}

\item{n}{the total number of respnodents in the proposed sample}
}
\description{
This function defines a model of careless responder. Objects using this class
are initialized with the following parameters
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{data}
\keyword{model}
