% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomGaussian.R
\name{getRandomGaussian}
\alias{getRandomGaussian}
\title{Choose a random value from a Gaussian distribution}
\usage{
getRandomGaussian(mu, sigma, max)
}
\arguments{
\item{mu}{the mean of the distribution}

\item{sigma}{the standard deviation of the distribution}
}
\description{
the size of careless populations in my simulations have a fairly small mean
with a reasonable standard deviation. In this and other scenarios it's
possible to select random values less than 0%, which doesn't have meaning.
If we were to round to zero then that would alter its shape. Thus we just
resample until we get a valid value
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
