% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridmap.R
\name{gridmap}
\alias{gridmap}
\title{A function for creating an arrangement of plots.}
\usage{
gridmap(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{a series of plot objects}

\item{plotlist}{optionally a list of plot objects}

\item{cols}{the number of columns for the grid}

\item{layout}{a layout for the grid object}
}
\description{
This takes a list of graph objects and arrangese them in a grid. I modified
a function I shamelessly took from stack overflow
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
