% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexOfMode.R
\name{indexOfMode}
\alias{indexOfMode}
\title{A function for finding the index of the mode of a dataset}
\usage{
indexOfMode(x)
}
\arguments{
\item{x}{an nx2 data frame}
}
\description{
assuming the group by variable is in the first column, this function
finds the index of the modal value of a data frame.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
