% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informedness.R
\name{informedness}
\alias{informedness}
\title{Implements an informedness class}
\usage{
informedness(range, informedness)
}
\arguments{
\item{range}{the range over which informedness was computed}

\item{informedness}{the valus for informedness}
}
\description{
This class returns an informedness object
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
