% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.equal.R
\name{is.equal}
\alias{is.equal}
\title{rowise vectorisation of all.equal}
\usage{
is.equal(x, y, ...)
}
\arguments{
\item{x}{a vector of values to compare to y}

\item{y}{a vector of values to compare to x}
}
\description{
This function returns a logical vector of all.equal evaluations
for each row of x and y
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
