% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jLookup.r
\name{jLookup}
\alias{jLookup}
\title{A function for finding y of x}
\usage{
jLookup(x, i)
}
\arguments{
\item{x}{an nx2 data frame}

\item{i}{index at which to report informedness}
}
\description{
This functiont takes a two column data frame as input. It
computes the max of the second column, and the informedness corresponding
to i
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
