% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labelCarelessClasses.R
\name{labelCarelessClasses}
\alias{labelCarelessClasses}
\title{A function for identifying careless classes using LPA}
\usage{
labelCarelessClasses(means, classes)
}
\arguments{
\item{x}{a datafreame of simulated data}
}
\description{
This function employs various rules of thumb that a person might use to
identify careless classes resulting from an LPA
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
