% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadpackages.R
\name{loadpackages}
\alias{loadpackages}
\title{A convenience function for loading packages}
\usage{
loadpackages(x = c("dplyr", "ggplot2", "grid", "gridExtra"))
}
\arguments{
\item{x}{a list of string package names}
}
\description{
A convenience function for loading packages
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{loading}
\keyword{package}
