% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longstring_classifier.R
\name{longstring_classifier}
\alias{longstring_classifier}
\title{Implements the longstring classifier evaluation}
\usage{
longstring_classifier(x)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe as an argument and evaluates the performance of longstring
as a classifier for that dataset.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
