% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longstringClassifier.R
\name{longstringClassifier}
\alias{longstringClassifier}
\title{Implements the longstring classifier evaluation}
\usage{
longstringClassifier(x, truth, cut = 0.4)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe as an argument and evaluates the performance of longstring
as a classifier for that dataset.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
