% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpaClassifier.R
\name{lpaClassifier}
\alias{lpaClassifier}
\title{A function for classifying careless responders using LPA}
\usage{
lpaClassifier(x)
}
\arguments{
\item{x}{a datafreame of simulated data}
}
\description{
yata yata yata
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
