% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahadClassifier.R
\name{mahadClassifier}
\alias{mahadClassifier}
\title{Implements the mahad classifier evaluation}
\usage{
mahadClassifier(x, cut = 0.5)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe as an argument and evaluates the performance of mahad
as a classifier for that dataset.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
