% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeRownames.R
\name{makeRownames}
\alias{makeRownames}
\title{Assigns rownames, because rbind.fill doesn't support these.}
\usage{
makeRownames(n)
}
\arguments{
\item{n}{a character list of rownames to assign}
}
\description{
Assigns rownames, because rbind.fill doesn't support these.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{loading}
\keyword{package}
