% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelSlopes.R
\name{modelSlopes}
\alias{modelSlopes}
\title{Extracts slopes from item parameters}
\usage{
modelSlopes(ipar, names, nchar)
}
\arguments{
\item{ipar}{matrix of item parameters from mirt}

\item{names}{the names of the items}

\item{nchar}{the number of characters to extract from the names.}
}
\description{
Given a matrix of item parameters that are generated from mirt this function
extracts the a parameter. This function is pretty specific to my dissertation
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
