% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Shamelessly copied and moded from stack overflow}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{any number of plot objects}

\item{cols}{the number of plots}

\item{plotList}{optionally, a list of plots}
}
\description{
Shamelessly copied and moded from stack overflow
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
