% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimizeMahaD.R
\name{optimizeMahaD}
\alias{optimizeMahaD}
\title{Implements the evenodd decision rule optimizer}
\usage{
optimizeMahaD(x, ...)
}
\arguments{
\item{x}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe .
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
