% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.crmetric.R
\name{plot.crmetric}
\alias{plot.crmetric}
\title{A method for plotting crmetric objects}
\usage{
\method{plot}{crmetric}(x, thisMetric, i)
}
\arguments{
\item{x}{an crmetric object}

\item{i}{a figure number for the graph}

\item{metric}{a label for the metric on the plot}
}
\description{
Takes an informedness object and plots it using ggplot
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
