% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.informedness.R
\name{plot.informedness}
\alias{plot.informedness}
\title{A method for plotting informedness objects}
\usage{
\method{plot}{informedness}(x, i)
}
\arguments{
\item{x}{an informedness object}

\item{i}{a figure number for the graph}
}
\description{
Takes an informedness object and plots it using ggplot
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
