% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBehavior.R
\name{plotBehavior}
\alias{plotBehavior}
\title{A method for plotting crmetric objects}
\usage{
plotBehavior(x, method, ...)
}
\arguments{
\item{x}{a df generated by runnign rq1}

\item{min}{the minmum of the interval over which to evaluate the behavior}

\item{max}{the maximum of the interval over which to evaluate the behavior}
}
\description{
Takes a df generated by RQ1 and plots the behavior of its metrics 
for the specified subsamples.
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
