% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recursiveAssign.R
\name{recursiveAssign}
\alias{recursiveAssign}
\title{Recursively assigns participants to condition}
\usage{
recursiveAssign(x, d, labels)
}
\arguments{
\item{x}{a matrix of survey responses}

\item{d}{a vector representing the number of records to assign to each label, i think.}

\item{a}{vector of labels}
}
\description{
Given a matrix of survey responses, a number of participants to select, and 
a vector of conditions to select them into, this function recursively assigns
participants at random.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
