% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse.R
\name{reverse}
\alias{reverse}
\title{Takes a vector or scalar and reverse codes it}
\usage{
reverse(var, max)
}
\arguments{
\item{var}{the value or vector of values to be recoded}

\item{max}{the highest value that var could theoretically have (e.g. 7 on a seven point likert scale)}
}
\description{
Takes a vector or scalar and reverse codes it
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
