% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverseCode.R
\name{reverseCode}
\alias{reverseCode}
\title{Reverse coding for an entire measure}
\usage{
reverseCode(x, neg, max)
}
\arguments{
\item{x}{a dataframe containing the construct items}

\item{neg}{a vector of integers denoting the negatively worded items}

\item{max}{the highest value that var could theoretically have (e.g. 7 on a seven point likert scale)}
}
\description{
Reverse coding for an entire measure
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
