% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmfiles.R
\name{rmfiles}
\alias{rmfiles}
\title{A convenience function for removing files in a directory}
\usage{
rmfiles(path)
}
\arguments{
\item{path}{the path to the files to be deleted}
}
\description{
A convenience function for removing files in a directory
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{loading}
\keyword{package}
