% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdInquisition.R
\name{sdInquisition}
\alias{sdInquisition}
\title{Implements the evenodd decision rule optimizer}
\usage{
sdInquisition(metric, truth, crModel, ...)
}
\arguments{
\item{metric}{a dataframe on which to test the classifier}

\item{truth}{a binary vector of classification labels}

\item{crModel}{a vector of labels for the CR model from which each CR was drawn}
}
\description{
Takes a dataframe .
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
