% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdPredictor.R
\name{sdPredictor}
\alias{sdPredictor}
\title{A function to calculate the CI of informedness}
\usage{
sdPredictor(x, cut)
}
\arguments{
\item{x}{a vector of values to provide predictions for}

\item{cut}{a scalar value representing the number of standard deviations to use as a cut score.}
}
\description{
This function takes a point estimate of informedness, a simulation size,
and a level of confidence as parameters.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confidence}
\keyword{interval}
