% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdRocer.R
\name{sdRocer}
\alias{sdRocer}
\title{Implements the evenodd decision rule optimizer}
\usage{
sdRocer(metric, truth, from, to, by)
}
\arguments{
\item{metric}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe .
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
