% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCareless.R
\name{simCareless}
\alias{simCareless}
\title{An interface for simulating careless respondents}
\usage{
simCareless(x, ...)
}
\arguments{
\item{x}{a vector of item responses}

\item{...}{any additional parameters needed to simulate a respondent of that type}

\item{type}{the label, or type of careless responder}
}
\description{
Given a matrix of item parameters that are generated from mirt this function
extracts the a parameter. This function is pretty specific to my dissertation
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
