% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simCenteredCareless.R
\name{simCenteredCareless}
\alias{simCenteredCareless}
\title{Simulates a centered careless respondent}
\usage{
simCenteredCareless(x, ...)
}
\arguments{
\item{x}{a dataframe containing the simulated response vector}

\item{floor}{the lowest possible value for each response}

\item{ceiling}{the highest possible value for each response}
}
\description{
Simulates a centered careless respondent
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
