% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFactorScores.R
\name{simFactorScores}
\alias{simFactorScores}
\title{A factor score simulator}
\usage{
simFactorScores(x, n)
}
\arguments{
\item{x}{a matrix of factor correlations}

\item{n}{the number of respondents to simulate}
}
\description{
This function simulates factor scores for n respondents such that the sample
has a factor structure approximating x, a matrix of factor correlations.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{factor}
\keyword{scores}
