% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLongstringCareless.R
\name{simLongstringCareless}
\alias{simLongstringCareless}
\title{Simulates a longstring careless respondent}
\usage{
simLongstringCareless(x, ...)
}
\arguments{
\item{x}{a dataframe containing the simulated response vector}

\item{floor}{the lowest possible value for each response}

\item{ceiling}{the highest possible value for each response}
}
\description{
Simulates a longstring careless respondent
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
