% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSkewedCareless.R
\name{simSkewedCareless}
\alias{simSkewedCareless}
\title{Simulates a skewed careless respondent}
\usage{
simSkewedCareless(x, ...)
}
\arguments{
\item{ipar}{matrix of item parameters from mirt}

\item{names}{the names of the items}

\item{nchar}{the number of characters to extract from the names.}
}
\description{
Given a vector of survey responses, an insert, and a number of items,
this function simulates a skewed careless respondent.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confusion}
\keyword{matrix}
