% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voteScore.R
\name{voteScore}
\alias{voteScore}
\title{Scores the votes of different methods of determining careless responders}
\usage{
voteScore(v)
}
\arguments{
\item{v}{a dataframe on which to test the classifier}
}
\description{
Takes a dataframe of votes as a parameter
}
\author{
Richard D. Yentes \email{rdyentes@ncsu.edu}
}
