% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informCI.R
\name{informCI}
\alias{informCI}
\title{A function to calculate the CI of informedness}
\usage{
informCI(b, n, alpha = 0.05)
}
\arguments{
\item{b}{the point estimate for informedness}

\item{n}{the number of observations in the sample}

\item{alpha}{the significance}
}
\description{
This function takes a point estimate of informedness, a simulation size,
and a level of confidence as parameters.
}
\author{
Richard D. Yentes \email{ryentes@ncsu.edu}
}
\keyword{confidence}
\keyword{interval}
